+++ geotrek/trekking/tests/test_parsers.py
@@ -9,6 +9,7 @@ from urllib.parse import urlparse
 
 from django.conf import settings
 from django.contrib.gis.geos import LineString, MultiLineString, Point, WKTWriter
+from django.core.exceptions import ImproperlyConfigured
 from django.core.management import call_command
 from django.core.management.base import CommandError
 from django.test import SimpleTestCase, TestCase
@@ -38,6 +39,7 @@ from geotrek.trekking.models import (
 )
 from geotrek.trekking.parsers import (
     ApidaePOIParser,
+    ApidaeServiceParser,
     ApidaeTrekParser,
     ApidaeTrekThemeParser,
     GeotrekPOIParser,
@@ -1963,6 +1965,79 @@ class ApidaePOIParserTests(TestCase):
         self.assertEqual(Attachment.objects.count(), 0)
 
 
+class TestApidaeServiceParser(ApidaeServiceParser):
+    url = "https://example.net/fake/api/"
+    api_key = "ABCDEF"
+    project_id = 1234
+    selection_id = 654321
+    service_type = "Foo"
+
+
+class TestApidaeServiceParserMissingType(ApidaeServiceParser):
+    url = "https://example.net/fake/api/"
+    api_key = "ABCDEF"
+    project_id = 1234
+    selection_id = 654321
+
+
+class ApidaeServiceParserTests(TestCase):
+    @staticmethod
+    def make_dummy_get(apidae_data_file):
+        return make_dummy_apidae_get(
+            parser_class=TestApidaeServiceParser,
+            test_data_dir="geotrek/trekking/tests/data/apidae_service_parser",
+            data_filename=apidae_data_file,
+        )
+
+    def test_no_service_type(self):
+        with self.assertRaisesMessage(
+            ImproperlyConfigured,
+            "A service type must be defined in parser configuration.",
+        ):
+            TestApidaeServiceParserMissingType()
+
+    @mock.patch("requests.get")
+    def test_skip_row_and_continue_when_wrong_geometry(self, mocked_get):
+        mocked_get.side_effect = self.make_dummy_get(
+            "services_incorrect_geometries.json"
+        )
+
+        output = StringIO()
+        call_command(
+            "import",
+            "geotrek.trekking.tests.test_parsers.TestApidaeServiceParser",
+            verbosity=2,
+            stdout=output,
+        )
+
+        self.assertEqual(Service.objects.count(), 0)
+        self.assertIn(
+            "Could not parse geometry from value '{'coordinates': [0, 0]}'",
+            output.getvalue(),
+        )
+        self.assertIn(
+            "Could not parse geometry from value '{'type': 'Point'}'",
+            output.getvalue(),
+        )
+
+    @mock.patch("requests.get")
+    def test_service_is_imported(self, mocked_get):
+        mocked_get.side_effect = self.make_dummy_get("service.json")
+
+        call_command(
+            "import",
+            "geotrek.trekking.tests.test_parsers.TestApidaeServiceParser",
+            verbosity=0,
+        )
+
+        self.assertEqual(Service.objects.count(), 1)
+        service = Service.objects.all().first()
+        self.assertEqual(service.type.name, "Foo")
+        self.assertEqual(service.eid, "1")
+        self.assertAlmostEqual(service.geom.coords[0], 813833.6, delta=0.1)
+        self.assertAlmostEqual(service.geom.coords[1], 6324255.0, delta=0.1)
+
+
 class PrepareAttachmentFromIllustrationTests(TestCase):
     def setUp(self):
         self.illustration = {
@@ -2410,6 +2485,7 @@ class OpenStreetMapPOIParser(TestCase):
     @classmethod
     def setUpTestData(cls):
         cls.type = POIType.objects.create(label="Test")
+        FileType.objects.create(type="Photographie")
         cls.path = PathFactory.create(
             geom=LineString((5.8394587, 44.6918860), (5.9527022, 44.7752786), srid=4326)
         )
@@ -2434,9 +2510,9 @@ class OpenStreetMapPOIParser(TestCase):
         self.assertEqual(self.objects.count(), 4)
 
     def test_POI_eid_filter_OSM(self):
-        poi_eid = self.objects.all().values_list("eid", flat=True)
-        self.assertListEqual(list(poi_eid), ["N1", "W2", "W3", "R4"])
-        self.assertNotEqual(poi_eid, ["1", "2", "3", "4"])
+        poi_eid = self.objects.all().order_by("eid").values_list("eid", flat=True)
+        self.assertListEqual(list(poi_eid), ["N1", "R4", "W2", "W3"])
+        self.assertNotEqual(list(poi_eid), ["1", "2", "3", "4"])
 
     def test_default_name(self):
         poi1 = self.objects.get(eid="N1")
@@ -2461,48 +2537,34 @@ class OpenStreetMapPOIParser(TestCase):
         self.assertAlmostEqual(poi.geom.x, 924596.692, places=2)
         self.assertAlmostEqual(poi.geom.y, 6412498.122, places=2)
 
-    @skipIf(
-        not settings.TREKKING_TOPOLOGY_ENABLED, "Test with dynamic segmentation only"
-    )
-    def test_topology_way(self):
+    def test_imported_way(self):
         poi = self.objects.get(eid="W2")
-        self.assertAlmostEqual(poi.topo_object.offset, -1401.037, places=2)
-        poi_path = poi.topo_object.paths.get()
-        self.assertEqual(poi_path, self.path)
-        self.assertEqual(poi.topo_object.kind, "POI")
-
-    def test_topology_way_no_dynamic_segmentation(self):
-        poi = self.objects.get(eid="W2")
+        if settings.TREKKING_TOPOLOGY_ENABLED:
+            self.assertAlmostEqual(poi.topo_object.offset, -1401.037, places=2)
+            poi_path = poi.topo_object.paths.get()
+            self.assertEqual(poi_path, self.path)
+            self.assertEqual(poi.topo_object.kind, "POI")
         self.assertAlmostEqual(poi.geom.x, 926882.120, places=2)
         self.assertAlmostEqual(poi.geom.y, 6403317.111, places=2)
 
-    @skipIf(
-        not settings.TREKKING_TOPOLOGY_ENABLED, "Test with dynamic segmentation only"
-    )
-    def test_topology_polygon(self):
+    def test_imported_object_polygon(self):
         poi = self.objects.get(eid="W3")
-        self.assertAlmostEqual(poi.topo_object.offset, -1398.870, places=2)
-        poi_path = poi.topo_object.paths.get()
-        self.assertEqual(poi_path, self.path)
-        self.assertEqual(poi.topo_object.kind, "POI")
+        if settings.TREKKING_TOPOLOGY_ENABLED:
+            self.assertAlmostEqual(poi.topo_object.offset, -1398.99, places=2)
+            poi_path = poi.topo_object.paths.get()
+            self.assertEqual(poi_path, self.path)
+            self.assertEqual(poi.topo_object.kind, "POI")
 
-    def test_topology_polygon_no_dynamic_segmentation(self):
-        poi = self.objects.get(eid="W3")
-        self.assertAlmostEqual(poi.geom.x, 933501.240, places=2)
-        self.assertAlmostEqual(poi.geom.y, 6410680.482, places=2)
+        self.assertAlmostEqual(poi.geom.x, 933496.557, places=2)
+        self.assertAlmostEqual(poi.geom.y, 6410675.089, places=2)
 
-    @skipIf(
-        not settings.TREKKING_TOPOLOGY_ENABLED, "Test with dynamic segmentation only"
-    )
-    def test_topology_relation(self):
+    def test_imported_object_relation(self):
         poi = self.objects.get(eid="R4")
-        self.assertAlmostEqual(poi.topo_object.offset, 2589.235, places=2)
-        poi_path = poi.topo_object.paths.get()
-        self.assertEqual(poi_path, self.path)
-        self.assertEqual(poi.topo_object.kind, "POI")
-
-    def test_topology_relation_no_dynamic_segmentation(self):
-        poi = self.objects.get(eid="R4")
+        if settings.TREKKING_TOPOLOGY_ENABLED:
+            self.assertAlmostEqual(poi.topo_object.offset, -2589.235, places=2)
+            poi_path = poi.topo_object.paths.get()
+            self.assertEqual(poi_path, self.path)
+            self.assertEqual(poi.topo_object.kind, "POI")
         self.assertAlmostEqual(poi.geom.x, 930902.933, places=2)
         self.assertAlmostEqual(poi.geom.y, 6406011.138, places=2)